fix(ish)ing tests
