+++ test/constants.js
@@ -1,0 +1,27 @@
+const ethers = require('ethers');
+
+const FIFTY = ethers.BigNumber.from('50000000000000000000');
+const FIFTY_PERCENT = ethers.BigNumber.from('500000000000000000');
+const ONE = ethers.BigNumber.from('1000000000000000000');
+const ONE_HUNDRED = ethers.BigNumber.from('100000000000000000000');
+const ONE_TENTH = ethers.BigNumber.from('100000000000000000');
+const ONE_THOUSAND = ethers.BigNumber.from('1000000000000000000000');
+const ONE_THOUSAND_ONE_HUNDRED = ethers.BigNumber.from('1100000000000000000000');
+const SIXTY_PERCENT = ethers.BigNumber.from('600000000000000000');
+const TEN = ethers.BigNumber.from('10000000000000000000');
+const THIRTY_FIVE_PERCENT = ethers.BigNumber.from('350000000000000000');
+const TWO_HUNDREDTHS = ethers.BigNumber.from('20000000000000000');
+
+module.exports = {
+  FIFTY_PERCENT,
+  FIFTY,
+  ONE_HUNDRED,
+  ONE_TENTH,
+  ONE_THOUSAND_ONE_HUNDRED,
+  ONE_THOUSAND,
+  ONE,
+  SIXTY_PERCENT,
+  TEN,
+  THIRTY_FIVE_PERCENT,
+  TWO_HUNDREDTHS,
+};
