+++ test/elasticDAOFactoryTest.js
@@ -1,14 +1,10 @@
 const { expect } = require('chai');
-const { SDK } = require('@elastic-dao/sdk');
 const ethers = require('ethers');
 const hre = require('hardhat').ethers;
-const { deployments } = require('hardhat');
-const { env } = require('./env');
+const SDK = require('@elastic-dao/sdk');
 
-const HUNDRED = ethers.BigNumber.from('100000000000000000000');
-const ONE = ethers.BigNumber.from('1000000000000000000');
-const ONE_TENTH = ethers.BigNumber.from('100000000000000000');
-const TWO_HUNDREDTHS = ethers.BigNumber.from('20000000000000000');
+const { ONE, ONE_HUNDRED, ONE_TENTH, TWO_HUNDREDTHS } = require('./constants');
+const env = require('./env');
 
 describe('ElasticDAO: Factory', () => {
   let agent;
@@ -18,16 +14,15 @@ describe('ElasticDAO: Factory', () => {
 
   it('Should allow a DAO to be deployed using the factory', async () => {
     [agent, summoner, summoner1, summoner2] = await hre.getSigners();
-    const provider = await hre.provider;
-    const sdk = new SDK({
+    const { provider } = hre;
+
+    const sdk = SDK({
       account: agent.address,
       contract: ({ abi, address }) => new ethers.Contract(address, abi, agent),
       env,
       provider,
       signer: agent,
     });
-
-    await deployments.fixture();
 
     const dao = await sdk.elasticDAOFactory.deployDAOAndToken(
       [summoner.address, summoner1.address, summoner2.address],
@@ -37,7 +32,7 @@ describe('ElasticDAO: Factory', () => {
       'EGT',
       ONE_TENTH,
       TWO_HUNDREDTHS,
-      HUNDRED,
+      ONE_HUNDRED,
       ONE,
     );
